graph: move `extension.rs` to `ext/futures.rs`
