+++ graph/src/ext/mod.rs
@@ -1,0 +1,2 @@
+///! Extension traits for external types.
+pub mod futures;
